refactoring the connection component
