+++ src/components/DBT/DBTCreateProfile.tsx
@@ -1,4 +1,4 @@
-import { Box, Button, CircularProgress, TextField } from '@mui/material';
+import { Box, Button, TextField } from '@mui/material';
 import { useForm } from 'react-hook-form';
 import { useSession } from 'next-auth/react';
 import { useContext, useState } from 'react';
